backend/requirements.txt
flask==2.3.3
flask-socketio==5.3.6
flask-cors==4.0.0
playwright==1.40.0
asyncio==3.4.3
python-socketio==5.9.0
eventlet==0.33.3
redis==5.0.1
celery==5.3.4
requests==2.31.0
python-dotenv==1.0.0